clean-up, restructuring and preparation for next studies
