+++ docker-compose/compose.local-test.yaml
@@ -1,0 +1,148 @@
+services:
+  kafka:
+    image: docker.io/bitnami/kafka:3.6.1@sha256:b5254050e61d1912ed9cf3f7107adc6a21ff29d85047b46d7b6df57cf25fa2f7
+    restart: unless-stopped
+    cap_drop:
+      - ALL
+    privileged: false
+    ipc: none
+    security_opt:
+      - "no-new-privileges:true"
+    environment:
+      KAFKA_CFG_NODE_ID: "0"
+      KAFKA_RAFT_CLUSTER_ID: "onco-analytics-on-fhir"
+      KAFKA_CFG_CONTROLLER_QUORUM_VOTERS: "0@kafka:9093"
+      KAFKA_CFG_PROCESS_ROLES: "controller,broker"
+      ALLOW_PLAINTEXT_LISTENER: "yes"
+      KAFKA_CFG_MESSAGE_MAX_BYTES: "31457280"
+      KAFKA_CFG_LISTENERS: "PLAINTEXT://:9092,CONTROLLER://:9093,EXTERNAL://:9094"
+      KAFKA_CFG_ADVERTISED_LISTENERS: "PLAINTEXT://kafka:9092,EXTERNAL://localhost:9094"
+      KAFKA_CFG_LISTENER_SECURITY_PROTOCOL_MAP: "CONTROLLER:PLAINTEXT,EXTERNAL:PLAINTEXT,PLAINTEXT:PLAINTEXT"
+      KAFKA_CFG_LOG_CLEANUP_POLICY: compact
+      KAFKA_CFG_CONTROLLER_LISTENER_NAMES: CONTROLLER
+      KAFKA_CFG_INTER_BROKER_LISTENER_NAME: PLAINTEXT
+    ports: # hier
+      - 127.0.0.1:9094:9094
+    volumes:
+      - "kafka_data:/bitnami:rw"
+
+  kafka-connect:
+    image: ghcr.io/miracum/util-images/cricketeerone-kafka-connect:v1.6.0@sha256:5d4c79d5b9d64562d8906aec14c66ed5f48ddb0f9e7f12949547ab89b69e3f0f
+    restart: unless-stopped
+    cap_drop:
+      - ALL
+    privileged: false
+    ipc: none
+    security_opt:
+      - "no-new-privileges:true"
+    environment:
+      CONNECT_BOOTSTRAP_SERVERS: kafka:9092
+      CONNECT_GROUP_ID: kafka-connect-group
+      CONNECT_KEY_CONVERTER: "org.apache.kafka.connect.storage.StringConverter"
+      CONNECT_VALUE_CONVERTER: "org.apache.kafka.connect.json.JsonConverter"
+      CONNECT_REST_ADVERTISED_HOST_NAME: "kafka-connect"
+      CONNECT_CONFIG_STORAGE_REPLICATION_FACTOR: "1"
+      CONNECT_OFFSET_STORAGE_REPLICATION_FACTOR: "1"
+      CONNECT_STATUS_STORAGE_REPLICATION_FACTOR: "1"
+      CONNECT_CONFIG_STORAGE_TOPIC: __kafka-connect-config
+      CONNECT_OFFSET_STORAGE_TOPIC: __kafka-connect-offsets
+      CONNECT_STATUS_STORAGE_TOPIC: __kafka-connect-status
+      CONNECT_PLUGIN_PATH: /app/libs
+    volumes:
+      - ./kafka-connect-passwords.properties:/tmp/kafka-connect-passwords.properties:ro
+    ports:
+      - 8083:8083
+
+  akhq:
+    image: docker.io/tchiotludo/akhq:0.24.0@sha256:6ccf8323ae6e93a893107f857cd9f7210add3569743b2c8528c6567967cc636f
+    restart: unless-stopped
+    cap_drop:
+      - ALL
+    privileged: false
+    ipc: none
+    security_opt:
+      - "no-new-privileges:true"
+    environment:
+      AKHQ_CONFIGURATION: |
+        akhq:
+          connections:
+            kafka:
+              properties:
+                bootstrap.servers: "kafka:9092"
+              connect:
+                - name: "kafka-connect"
+                  url: "http://kafka-connect:8083"
+    ports:
+      - "8084:8080"
+    depends_on:
+      - kafka
+
+  decompose-xmls:
+    image: ghcr.io/bzkf/onco-analytics-on-fhir/decompose-xmls:v2.3.1 # x-release-please-version
+    user: "${UID:-1000}:${GID:-1000}"
+    environment:
+      KAFKA_ENABLED: "true"
+      KAFKA_BOOTSTRAP_SERVERS: "kafka:9092"
+      KAFKA_OUTPUT_TOPIC: "onkostar.MELDUNG_EXPORT"
+      SAVE_AS_FILES_ENABLED: "false"
+      # Set to 'true' to enable removing of leading zeros in patient IDs
+      REMOVE_LEADING_PATIENTID_ZEROS: "false"
+    volumes:
+      - ./input-obds-reports:/app/input-obds-reports:ro
+      - ./output-obds-reports:/app/output-obds-reports:rw
+
+  obds-to-fhir:
+    image: ghcr.io/bzkf/obds-to-fhir:v2.2.0@sha256:b5d264c849fb5a19d8af1677b6f9adc26d70108c22e4f2753d0b2472da39f038
+    restart: unless-stopped
+    cap_drop:
+      - ALL
+    privileged: false
+    ipc: none
+    security_opt:
+      - "no-new-privileges:true"
+    user: "65532:65532"
+    environment:
+      SECURITY_PROTOCOL: PLAINTEXT
+      BOOTSTRAP_SERVERS: kafka:9092
+      FHIR_OUTPUT_TOPIC_PARTITION_COUNT: "1"
+      SPRING_CLOUD_STREAM_KAFKA_BINDER_MINPARTITIONCOUNT: "1"
+      SPRING_CLOUD_STREAM_KAFKA_BINDER_AUTOADDPARTITIONS: "true"
+      NUM_STREAM_THREADS: "1"
+      JAVA_TOOL_OPTIONS: "-XX:MaxRAMPercentage=75"
+      ACTIVE_PROFILE: "patient"
+      YEAR_LIMITATION_SUFFIX: "" # e.g. ".2023"
+      FHIR_SYSTEMS_PATIENTID: ${FHIR_SYSTEMS_PATIENTID:-https://fhir.diz.uk-erlangen.de/identifiers/patient-id}
+      FHIR_SYSTEMS_CONDITIONID: ${FHIR_SYSTEMS_CONDITIONID:-https://fhir.diz.uk-erlangen.de/identifiers/onkostar-xml-condition-id}
+      FHIR_SYSTEMS_OBSERVATIONID: ${FHIR_SYSTEMS_OBSERVATIONID:-https://fhir.diz.uk-erlangen.de/identifiers/onkostar-xml-observation-id}
+      FHIR_SYSTEMS_PROCEDUREID: ${FHIR_SYSTEMS_PROCEDUREID:-https://fhir.diz.uk-erlangen.de/identifiers/onkostar-xml-procedure-id}
+      FHIR_SYSTEMS_MEDICATIONSTATEMENTID: ${FHIR_SYSTEMS_MEDICATIONSTATEMENTID:-https://fhir.diz.uk-erlangen.de/identifiers/onkostar-xml-medication-statement-id}
+    depends_on:
+      - decompose-xmls
+
+  obds-fhir-to-opal:
+    image: ghcr.io/bzkf/onco-analytics-on-fhir/obds-fhir-to-opal:v2.3.1 # x-release-please-version
+    restart: no
+    cap_drop:
+      - ALL
+    privileged: false
+    ipc: private
+    security_opt:
+      - "no-new-privileges:true"
+    user: "1001:1001"
+    environment:
+      OUTPUT_FOLDER: "/home/spark/opal-output"
+      KAFKA_TOPIC_YEAR_SUFFIX: "" # e.g. ".2023"
+      KAFKA_BOOTSTRAP_SERVER: "kafka:9092"
+      KAFKA_PATIENT_TOPIC: "fhir.obds.Patient"
+      KAFKA_CONDITION_TOPIC: "fhir.obds.Condition"
+      KAFKA_OBSERVATION_TOPIC: "fhir.obds.Observation"
+      KAFKA_PROCEDURE_TOPIC: "fhir.obds.Procedure"
+      KAFKA_MEDICATIONSTATEMENT_TOPIC: "fhir.obds.MedicationStatement"
+    volumes:
+      - ./opal-output:/home/spark/opal-output:rw
+    depends_on:
+      - obds-to-fhir
+
+volumes:
+  kafka_data:
+    driver: local
