+++ src/obds_fhir_to_opal/README.md
@@ -6,7 +6,14 @@
 
 - **Convert FHIR resources from Kafka topics** into a structured CSV file.
 - **Generate a data dictionary** for the Federated Analysis with DataSHIELD.
-- **Result CSV File Columns**:
+- **Data Extraction for specific studies** can be controlled via an environment variable `STUDY_NAME` in the compose.yaml file.
+`STUDY_NAME:`
+  - `poc`: Proof of Concept Data Extraction used in our publication https://preprints.jmir.org/preprint/65681
+  - `study_protocol_a`: data extraction script for the analysis part A from our BZKF study protocol (latest amendment from 09.12.2024)
+  - `study_protocol_c`: data extraction script for the analysis part C from our BZKF study protocol (latest amendment from 09.12.2024)
+
+### File Structure of the data extraction results
+- `poc`:
   - `ID`: Unique identifier for the record.
   - `condition_id`: Identifier for the medical condition.
   - `date_diagnosis`: Date when the diagnosis was made.
@@ -18,14 +25,27 @@
   - `icd10_mapped`: Mapped ICD-10 code.
   - `icd10_grouped_entities`: Grouped ICD-10 entities.
   - `gender_mapped`: Mapped gender information.
+- `study_protocol_a`:
+  - ...
+- `study_protocol_c`:
+  - `ID`: Unique identifier for the record.
+  - `condition_id`: Identifier for the medical condition.
+  - `date_diagnosis`: Date when the diagnosis was made.
+  - `icd10_code`: ICD-10 code for the diagnosis.
+  - `gender`: Gender of the patient.
+  - `date_diagnosis_year`: Year when the diagnosis was made.
+  - `icd10_mapped`: Mapped ICD-10 code.
+  - `icd10_grouped`: ICD-10 code grouped to parent code.
+  - `icd10_entity`: Grouped ICD-10 entities.
+  - `gender_mapped`: Mapped gender information.
 
-  For a more detailed description of the data elements, please check the data_dictionary_df.xlsx.
+  For a more detailed description of the data elements, please check each data_dictionary_df.xlsx.
 
 ## DataSHIELD R Scripts
 
-In addition to the main tool, this repository contains a subfolder `DataSHIELD_R-Script`. This folder includes R scripts for a Federated Analysis querying the result CSV file with help of DataSHIELD.
+In addition to the main tool, this repository contains a subfolder `DataSHIELD_R-Script`. This folder contains R scripts for a Federated Analysis querying the result CSV file with help of DataSHIELD.
 
-Results from this analysis will be published soon. Check back for updates.
+Results from this analysis will be published soon, Preprint available here: https://preprints.jmir.org/preprint/65681
 
 ## Contributing
