+++ src/obds_fhir_to_opal/compose.yaml
@@ -6,7 +6,8 @@ services:
     environment:
       OUTPUT_FOLDER: "/opt/bitnami/spark/opal-output/"
       OUTPUT_FILENAME: "df.csv"
-      KAFKA_TOPIC_YEAR_SUFFIX: ".2022"
+      STUDY_NAME: "PoC"
+      KAFKA_TOPIC_YEAR_SUFFIX: "" #".2022"
       KAFKA_BOOTSTRAP_SERVER: "kafka:9092"
       KAFKA_PATIENT_TOPIC: "fhir.obds.Patient"
       KAFKA_CONDITION_TOPIC: "fhir.obds.Condition"
@@ -14,4 +15,4 @@ services:
       KAFKA_PROCEDURE_TOPIC: "fhir.obds.Procedure"
       KAFKA_MEDICATIONSTATEMENT_TOPIC: "fhir.obds.MedicationStatement"
     volumes:
-      - ${PWD}/output:/home/spark/opal-output
+      - ./output:/opt/bitnami/spark/opal-output
