+++ src/obds_fhir_to_opal/output/PoC/df.csv
@@ -1,0 +1,3 @@
+condition_id,date_diagnosis,icd10_code,gender,date_diagnosis_year,date_diagnosis_month,date_diagnosis_day,icd10_mapped,icd10_grouped_entities,gender_mapped,ID
+69778e3e81986f370848ef0a36501db7dbff9faaa80315a59bf2aeea9a98906a,1900-07-29,C61,male,1900,7,29,361.0,14,2,0
+4d894939aabb8883d54f1193ec273f1202e142b89134106b2b41310db2059d3f,1900-08-08,D07.3,female,1900,8,8,407.3,-100,1,1
