+++ src/obds_fhir_to_opal/utils_onco_analytics.py
@@ -1,8 +1,14 @@
 from typing import List, Optional
+import os
+import shutil
 
 import pandas as pd
+from pathling import PathlingContext, datasource
+from pathling import Expression as exp
+
 from pyspark.sql import DataFrame
-from pyspark.sql import functions as F
+from pyspark.sql.functions import monotonically_increasing_id, udf, year, month, dayofmonth
+from pyspark.sql.types import DoubleType, IntegerType, StringType
 
 
 def map_icd10(icd10_code):
@@ -108,11 +114,114 @@ def deconstruct_date(df: DataFrame, date_col: str):
     month_col = f"{date_col}_month"
     day_col = f"{date_col}_day"
 
-    df = df.withColumn(year_col, F.year(date_col))
-    df = df.withColumn(month_col, F.month(date_col))
-    df = df.withColumn(day_col, F.dayofmonth(date_col))
+    df = df.withColumn(year_col, year(date_col))
+    df = df.withColumn(month_col, month(date_col))
+    df = df.withColumn(day_col, dayofmonth(date_col))
 
     return df
+
+
+def lookup_gender(gender_string):
+    if gender_string is None or gender_string == "":
+        gender_mapped = 0
+    elif gender_string == "female" or gender_string == "weiblich":
+        gender_mapped = 1
+    elif gender_string == "male" or gender_string == "männlich":
+        gender_mapped = 2
+    else:
+        gender_mapped = 3  # other / divers
+    return gender_mapped
+
+
+def return_year(deceasedDateTime):
+    if deceasedDateTime is not None:
+        year = deceasedDateTime[0:4]
+        return year
+    else:
+        return deceasedDateTime
+
+
+def extract_df_PoC(pc: PathlingContext, data: datasource.DataSource):
+    map_icd10UDF = udf(lambda x: map_icd10(x), StringType())
+    group_entitiesUDF = udf(lambda x: group_entities(x), StringType())
+    map_genderUDF = udf(lambda x: map_gender(x), StringType())
+
+    df = data.extract(
+        "Condition",
+        columns=[
+            exp("id", "condition_id"),
+            exp("onsetDateTime", "date_diagnosis"),
+            exp(
+                """code.coding
+                .where(system='http://fhir.de/CodeSystem/bfarm/icd-10-gm').code""",
+                "icd10_code",
+            ),
+            exp("Condition.subject.resolve().ofType(Patient).gender", "gender"),
+        ],
+    )
+    # deconstruct date in YYYY, MM, DD
+    df = deconstruct_date(df, "date_diagnosis")
+
+    # map ICD10 code to numeric value for DataSHIELD, C = 3, D = 4
+    df = df.withColumn(
+        "icd10_mapped",
+        map_icd10UDF(df.icd10_code).cast(DoubleType()),
+    )
+    # groupd icd10 to entities, inspired from "Jahresbericht 2023 des Bayerischen
+    # Krebsregisters - Krebs in Bayern in den Jahren 2015 bis 2019 - Band 5"
+    df = df.withColumn(
+        "icd10_grouped_entities",
+        group_entitiesUDF(df.icd10_mapped).cast(IntegerType()),
+    )
+    # map gender to numeric value for DataSHIELD, None = 0, "female" = 1, "male" = 2,
+    # "other/diverse" = 3
+    df = df.withColumn(
+        "gender_mapped",
+        map_genderUDF(df.gender).cast(IntegerType()),
+    )
+    df = df.dropDuplicates()
+
+    return df
+
+
+def save_final_df(df, settings):
+    df_with_id = df.withColumn("ID", monotonically_increasing_id())
+
+    df_with_id = df_with_id.coalesce(1)
+    output_dir = os.path.join(settings.output_folder, settings.study_name)
+    temp_dir = os.path.join(settings.output_folder, settings.study_name, "temp")
+
+    os.makedirs(temp_dir, exist_ok=True)
+    df_with_id.write.mode("overwrite").csv(temp_dir, header=True)
+    output_file = os.path.join(output_dir, settings.output_filename)
+    part_file = [f for f in os.listdir(temp_dir) if f.startswith("part-")][0]
+    shutil.move(os.path.join(temp_dir, part_file), output_file)
+    shutil.rmtree(temp_dir)
+
+
+def prepare_datadictionary_PoC(df):
+    # generate data dictionary
+    dtypes_list_df = [str(dtype[1]) for dtype in df.dtypes]
+
+    description_list_df_dictionary = {
+        "condition_id": "Condition ID, unique for each condition",
+        "date_diagnosis": "date of diagnosis",
+        "icd10_code": "ICD10 GM diagnosis code",
+        "icd10_mapped": """ICD10 GM diagnosis code mapped A = 1, B = 2, C = 3, D = 4,
+        e.g.: A01.9 = 101.9, C50.1 = 350.1 or D41.9 = 441.9""",
+        "icd10_grouped_entities": """ICD10 GM diagnosis code grouped to entity groups
+        from 0-23 according to LGL Report Cancer in Bavaria 2019, mapping see
+        github.com/bzkf/onco-analytics-on-fhir/src/obds_fhir_to_opal
+        /utils_onco_analytics.py""",
+        "date_diagnosis_year": "Year of Diagnosis",
+        "date_diagnosis_month": "Month of Diagnosis",
+        "date_diagnosis_day": "Day of Diagnosis",
+        "gender": "Gender - male, female, other/diverse",
+        "gender_mapped": """Gender mapped: 0 = None, 1 = female, 2 = male,
+        3 = other/diverse""",
+    }
+
+    return dtypes_list_df, description_list_df_dictionary
 
 
 # optionally pass lists to fill the datadictionary
@@ -180,6 +289,9 @@ def generate_datadictionary(
     # create second (empty) sheet for data dictionary "Categories"
     categories_columns = ["table", "variable", "name", "code", "missing"]
     df_categories = pd.DataFrame(columns=categories_columns)
+
+    parent_dir = os.path.dirname(file_path)
+    os.makedirs(parent_dir, exist_ok=True)
 
     with pd.ExcelWriter(file_path, engine="xlsxwriter") as writer:
         df.to_excel(writer, index=False, sheet_name="Variables")
